+++ ios/Classes/QrscanPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface QrscanPlugin : NSObject<FlutterPlugin>
-@end